Fix column types

+++ database/migrations/2024_08_14_170301_create_transaction_inputs_table.php
@@ -11,11 +11,11 @@ return new class extends Migration
         Schema::create('transaction_inputs', function (Blueprint $table) {
             $table->id();
             $table->string('tx_id', 64);
-            $table->unsignedTinyInteger('input_index');
+            $table->unsignedSmallInteger('input_index');
             $table->string('prev_tx_id', 64)->nullable();
             $table->unsignedInteger('prev_vout')->nullable();
             $table->string('address', 35)->nullable()->index();
-            $table->decimal('amount', 16, 8)->nullable();
+            $table->decimal('amount', 30, 8)->nullable();
             $table->text('script_sig')->nullable();
             $table->unsignedBigInteger('sequence')->default(4294967295);
             $table->text('coinbase_data')->nullable();
